flake8 errors and some refactoring
