+++ etl/lib/cloud_management.py
@@ -8,7 +8,11 @@ from google.cloud import storage
 
 def authenticate():
     return storage.Client.from_service_account_json(
+        # for running on the airflow GPU
         '/home/lukezhu/elvo-analysis/credentials/client_secret.json'
+
+        # for running locally
+        # 'credentials/client_secret.json'
     )
