+++ etl/lib/multichan_failure_analysis.py
@@ -1,6 +1,7 @@
 """
-Purpose: Manually analyze the false negative cases generated by our current best
-Resnet-50 model. Images are generate using the same code as multichannel_mip.py.
+Purpose: Manually analyze the false negative cases generated by our
+current best Resnet-50 model. Images are generated using the same
+code as multichannel_mip.py.
 """
 
 import logging
