+++ etl/mip.py
@@ -6,7 +6,7 @@ down into a single 2D array.
 
 # TODO: preprocess coronal and sagittal scans so they have mips too
 import logging
-# from matplotlib import pyplot as plt
+from matplotlib import pyplot as plt
 from lib import transforms, cloud_management as cloud
 
 WHENCE = ['numpy/axial',
@@ -42,7 +42,7 @@ def normal_mip():
             logging.info(f'downloading {in_blob.name}')
             input_arr = cloud.download_array(in_blob)
             logging.info(f"blob shape: {input_arr.shape}")
-            if location == 'numpy/axial':
+            if location == 'airflow/npy':
                 cropped_arr = transforms.crop_normal_axial(input_arr,
                                                            location)
             else:
@@ -51,9 +51,9 @@ def normal_mip():
             not_extreme_arr = transforms.remove_extremes(cropped_arr)
             logging.info(f'removed array extremes')
             mip_arr = transforms.mip_normal(not_extreme_arr)
-            # plt.figure(figsize=(6, 6))
-            # plt.imshow(mip_arr, interpolation='none')
-            # plt.show()
+            plt.figure(figsize=(6, 6))
+            plt.imshow(mip_arr, interpolation='none')
+            plt.show()
 
             # # if the source directory is one of the luke ones
             # if location != 'numpy':
